requirements.txt
absl-py==0.7.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gast==0.2.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.19.0
google-pasta==0.1.7
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.15.4
olefile==0.46
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
scipy==1.3.1
six==1.12.0
tbb==2019.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2